% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGMT.R
\name{prepGMT}
\alias{prepGMT}
\alias{prepGMT.data.frame}
\alias{prepGMT.character}
\title{prepGMT}
\usage{
prepGMT(pathway_list, ...)

\method{prepGMT}{data.frame}(pathway_list)

\method{prepGMT}{character}(pathway_list)
}
\arguments{
\item{pathway_list}{A data.frame consisting of two columns (`ont`` and `gene`) or the
path to a gmt file}

\item{...}{Not used}
}
\description{
Prepare a GMT file for use in score_module_eigengenes
}
\examples{

}
