% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengene_scoring.R
\name{score_matrix}
\alias{score_matrix}
\title{score_matrix}
\usage{
score_matrix(object, module_list, score_func = "rsvd")
}
\arguments{
\item{object}{matrix of gene expression to assign an eigenvalue}

\item{module_list}{list of gene modules}

\item{score_func}{function to use to assign an eigenvalue. Functions allowed include 'svd', 'rsvd', and 'nmf'}
}
\description{
score_matrix
}
