% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengene_scoring.R
\name{score_module_eigengenes}
\alias{score_module_eigengenes}
\alias{score_module_eigengenes.default}
\alias{score_module_eigengenes.DESeqDataSet}
\alias{score_module_eigengenes.Seurat}
\title{score_module_eigengenes}
\usage{
score_module_eigengenes(object, ...)

\method{score_module_eigengenes}{default}(object, module_list, md = NULL)

\method{score_module_eigengenes}{DESeqDataSet}(object, module_list)

\method{score_module_eigengenes}{Seurat}(object, module_list,
  assay = "RNA", slot = "data")
}
\arguments{
\item{object}{Either a matrix of expression values of a DESeqDataSet}

\item{module_list}{A named lists of lists where each sublist are the genes that make up the
module and each sublist is named.}

\item{md}{Optional meta data to add to the scores (for plotting purposes). Default: NULL}

\item{assay}{Seurat assay object from which to pull data}

\item{slot}{Assay slot to use}
}
\description{
score_module_eigengenes
}
